+++ SocketNetworking.Modding/InstanceFieldChangeSync.cs
@@ -103,7 +103,7 @@ namespace SocketNetworking.Modding
 
         public int GetLength()
         {
-            return Serialize().Length;
+            return (int)Serialize().Length;
         }
 
         public ByteWriter Serialize()
