+++ SocketNetworking/Shared/Authentication/AuthenticationProvider.cs
@@ -75,7 +75,7 @@ namespace SocketNetworking.Shared.Authentication
 
         public int GetLength()
         {
-            return Serialize().Length;
+            return (int)Serialize().Length;
         }
 
         public ByteWriter Serialize()
